Added IAuthorRepository to Dependency Injection container, and as a constructor parameter to ChirpRepository
